# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lobsense IoT V2 API Documentation",
        "description": "Dokumentasi resmi API sistem monitoring, telemetri sensor tambak, pengontrolan aerator/feeder, inferensi AI, dan manajemen perangkat Lobsense V2.",
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Lobsense API Local Server"
        }
    ],
    "paths": {
...
        "/api/v2/cameras": {
            "get": {
                "tags": [
                    "Cameras"
                ],
                "summary": "Daftar Semua Kamera",
                "description": "Mengambil seluruh data kamera pengawas yang terdaftar beserta relasi IoT Node & Cage.",
                "operationId": "b6a6a99c97e4050d3f3afcb4520d1a63",
                "responses": {
                    "200": {
                        "description": "Daftar kamera berhasil diambil"
                    },
                    "401": {
                        "description": "Tidak terautentikasi"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cameras"
                ],
                "summary": "Tambah Kamera Baru",
                "description": "Mendaftarkan kamera pengawas baru dan menghubungkannya ke IoT Node.",
                "operationId": "55bb27e31e98d1fbe8b60e47c8bc52ac",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "camera_code",
                                    "iot_node_id"
                                ],
                                "properties": {
                                    "camera_code": {
                                        "type": "string",
                                        "example": "CAM-A01"
                                    },
                                    "iot_node_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "stream_url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "rtsp://192.168.1.100:554/stream"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Kamera berhasil ditambahkan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/cameras/{id}": {
            "get": {
                "tags": [
                    "Cameras"
                ],
                "summary": "Detail Kamera",
                "description": "Mengambil detail data satu kamera berdasarkan ID.",
                "operationId": "66584f6784071b7489e8ac91f84a955a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail kamera berhasil diambil"
                    },
                    "404": {
                        "description": "Kamera tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Cameras"
                ],
                "summary": "Update Kamera",
                "description": "Memperbarui data kamera pengawas yang sudah ada.",
                "operationId": "dd3917d9e33e0f3127e48d5864a086aa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "camera_code": {
                                        "type": "string",
                                        "example": "CAM-B02"
                                    },
                                    "iot_node_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "stream_url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "rtsp://192.168.1.101:554/stream"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kamera berhasil diperbarui"
                    },
                    "404": {
                        "description": "Kamera tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cameras"
                ],
                "summary": "Hapus Kamera",
                "description": "Menghapus data kamera pengawas berdasarkan ID.",
                "operationId": "f2dfedefed8a50450858935cdba47776",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kamera berhasil dihapus"
                    },
                    "404": {
                        "description": "Kamera tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}